+++ hub-service/src/main/java/com/sparta/hubservice/hub/application/dto/response/GeocodeResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.hubservice.hub.application.dto;
+package com.sparta.hubservice.hub.application.dto.response;
 
 import java.util.List;
 import lombok.Data;
