+++ hub-service/src/main/java/com/sparta/hubservice/hub/application/dto/response/HubResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.hubservice.hub.application.dto;
+package com.sparta.hubservice.hub.application.dto.response;
 
 import com.sparta.hubservice.hub.domain.model.Hub;
 import java.math.BigDecimal;
@@ -7,31 +7,32 @@ import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @Builder
 @AllArgsConstructor
-public class HubCreateResponseDto {
+public class HubResponseDto {
 
     private final UUID hubId;
-    private final String message;
     private final String name;
     private final String address;
-    private final BigDecimal latitud;
-    private final BigDecimal longitud;
+    private final BigDecimal latitude;
+    private final BigDecimal longitude;
     private final LocalDateTime createdAt;
     private final long createdBy;
+    private final LocalDateTime updatedAt;
+    private final long updatedBy;
 
-    public HubCreateResponseDto(Hub hub, String message) {
+    public HubResponseDto(Hub hub) {
         this.hubId = hub.getHubId();
-        this.message = message;
         this.name = hub.getName();
         this.address = hub.getAddress();
-        this.latitud = hub.getLatitude();
-        this.longitud = hub.getLongitude();
-        this.createdAt = LocalDateTime.now();
+        this.latitude = hub.getLatitude();
+        this.longitude = hub.getLongitude();
+        this.createdAt = hub.getCreatedAt();
         this.createdBy = hub.getCreatedBy();
+        this.updatedAt = hub.getUpdatedAt();
+        this.updatedBy = hub.getUpdatedBy();
     }
 
 }
